Consolidate VHPI code
We had many copies of the VHPI marshalling/unmarshalling code.
